requirements.txt
#discord
python-dotenv
discord.py-self
openai
#pydantic
smolagents